requirements.txt
# Inline dependencies for Mini Pokedex CLI
requests==2.31.0
pymongo==4.7.2
python-dotenv==1.0.1
webexteamssdk==1.7
